+++ PC2-3/database/migrations/2021_04_26_114912_create_restaurantes_table.php
@@ -15,7 +15,7 @@ class CreateRestaurantesTable extends Migration
     {
         Schema::create('restaurantes', function (Blueprint $table) {
             $table->id();
-            $table->integer('idMunicipio');
+            $table->unsignedBigInteger('idMunicipio');
             $table->string('Nombre');
             $table->string('Detalles');
             $table->string('Comentario');
