+++ PC2-3/database/migrations/2021_04_26_114954_create_ocios_table.php
@@ -15,7 +15,7 @@ class CreateOciosTable extends Migration
     {
         Schema::create('ocios', function (Blueprint $table) {
             $table->id();
-            $table->integer('idMunicipio');
+            $table->unsignedBigInteger('idMunicipio');
             $table->string('Nombre');
             $table->string('Comentario');
             $table->string('Referencia');
